requirements.txt
Django
djangorestframework
markdown
django-filter
Pillow
psycopg2
celery
redis